final cleanup, rm comments, formatted files
