+++ dl_models/bert.py
@@ -20,7 +20,6 @@ class BERTClassifier(nn.Module):
         """
         input_ids, token_type_ids, attn_mask = inputs.permute(1, 0, 2)
         _, pooled_output = self.bert(input_ids, token_type_ids, attn_mask)
-        # _, pooled_output, attn_layers = self.bert(input_ids, token_type_ids, attn_mask)
         # dense 0 + relu
         output = self.activation_func(self.dense_0(pooled_output))
         # dropout
