+++ dl_models/dataset_classes.py
@@ -1,0 +1,27 @@
+import numpy as np
+import torch
+from torch.utils.data import Dataset
+
+
+class BERTDataset(Dataset):
+    def __init__(self, X, Y):
+        self.X = X
+        self.y = Y
+
+    def __len__(self):
+        return len(self.y)
+
+    def __getitem__(self, idx):
+        return self.X[idx], self.y.iloc[idx], 0  # length not needed for BERT
+
+
+class ValDataset(Dataset):
+    def __init__(self, X, Y):
+        self.X = X
+        self.y = Y
+
+    def __len__(self):
+        return len(self.y)
+
+    def __getitem__(self, idx):
+        return torch.from_numpy(self.X.iloc[idx][0].astype(np.int32)), self.y.iloc[idx], self.X.iloc[idx][1]
